requirements_backup.txt
altair==5.5.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
dash==3.0.3
dnspython==2.7.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
Faker==37.1.0
Flask==3.0.3
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
narwhals==1.35.0
nest-asyncio==1.6.0
numpy==2.0.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
plotly==6.0.1
protobuf==5.29.4
pyarrow==20.0.0
pydeck==0.9.1
pymongo==4.12.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
referencing==0.36.2
reportlab==4.4.0
requests==2.32.3
retrying==1.3.4
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.44.1
telegram==0.0.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.0.6
zipp==3.21.0
